Made a react specific copy button and refactored copyAddress function
