+++ src/html.jsx
@@ -1,4 +1,4 @@
-const { copyAddress } = require('./utils')
+const { CopyButton } = require('./CopyButton')
 const style = require('style-loader!./style.less')
 
 module.exports = ({ dom, React, Fragment }) => {
@@ -40,10 +40,6 @@ module.exports = ({ dom, React, Fragment }) => {
   )
 
   const AddressBarTemplate = ({ data = {}, opts = {} }) => {
-    const onClick = React
-      ? ({ target }) => copyAddress(target, data.address)
-      : `boxCopyAddress_f1kx(this, '${data.address}')`
-
     return (
       <div className={`${style.boxAddress} ${data.addressDisplay.length >= 15 ? style.boxAddressFull : ''}`}>
         <div className={style.boxImg}>
@@ -54,10 +50,7 @@ module.exports = ({ dom, React, Fragment }) => {
           {data.addressDisplay}
         </div>
 
-        <div className={style.addressCopy} onClick={onClick}>
-          <i className="far fa-clone clone"></i>
-          <i className="fas fa-check check" style={{display: 'none'}}></i>
-        </div>
+        <CopyButton address={data.address} />
         {opts.empty ? <EmptyHoverTemplate data={data} /> : <HoverTemplate data={data} opts={opts} />}
       </div>
     )
